+++ spec/spec_helper.rb
@@ -10,4 +10,11 @@ RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
 
   config.backtrace_clean_patterns = [/rspec/]
+
+  config.before do
+    @current_datetime = DateTime.now
+    DateTime.stub now: @current_datetime
+
+    Punchblock.stub new_request_id: 'foo'
+  end
 end